Introduce Postgres queue (#131)
* Introduce Postgres queue

* Add psycopg dependency

* Remove redis from install_requires

* Use AsyncConnectionPool

* Quote channel

* Clean up jsonb serialize/deserialize

* PR feedback

* Make jobs table configurable

* Support abort and sweep

* Add postgres service to CI

* Set CI postgres auth to trust

* Revert sweep

* WIP

* Add index for dequeue

* WIP

* Move uuid to base queue

* Add and fix tests

* Update abort test

* PR feedback

* Add sweep test

* Refactor dequeue

* Add sweeping

* Fix listen

* Re-enabled more tests

* Finish abort

* Refactor repetitive code

* Escape jobs table name

* Make enqueue channel name a constant

* PR feedback

* Rename lock_id to lock_key

* Add sweep lock

* Add pg to benchmark script

* Use same connection to notify enqueues

* Lock rows when sweeping

* PR feedback

* Write stats to database

* Add worker and job info

* Update tests

* WIP optimize dequeue

* Optimize release job

* PR feedback

* Fix mypy issue with tests

* PR feedback

* Sweep expired stats

* Increase listen timeout slightly in TestPostgresQueue.test_update

* Increase listen timeout slightly more in TestPostgresQueue.test_update

* Secure LISTEN and NOTIFY queries

* Revert timeout

* Add job tests for Postgres queue

* Test

* Undo

* Debugging

* Debugging

* Remove debugging statements

* Fix sweep query

* Mark task done when finishing a job

* Fix tests

* Update benchmarks

* Add lock test

* Remove _connected

* Format sql

* PR feedback

* PR feedback

* Refactor locks

* Use test_saq schema for tests

* Format

* Setup and teardown test_saq schema

* Update deps make target

* PR feedback

* Fix typos

* Refactor some base queue class functions

* Change job column to BYTEA

* Fix annotation

+++ tests/helpers.py
@@ -1,13 +1,47 @@
 import typing as t
 
-from saq.queue import Queue, RedisQueue
+import psycopg
+
+from saq.queue import Queue
+from saq.queue.postgres import PostgresQueue
+from saq.queue.redis import RedisQueue
+
+POSTGRES_TEST_SCHEMA = "test_saq"
 
 
-def create_queue(**kwargs: t.Any) -> RedisQueue:
+async def create_redis_queue(**kwargs: t.Any) -> RedisQueue:
     queue = t.cast(RedisQueue, Queue.from_url("redis://localhost:6379", **kwargs))
+    await queue.connect()
     return queue
 
 
-async def cleanup_queue(queue: RedisQueue) -> None:
-    await queue.redis.flushdb()
+async def create_postgres_queue(**kwargs: t.Any) -> PostgresQueue:
+    queue = t.cast(
+        PostgresQueue,
+        Queue.from_url(
+            f"postgres://postgres@localhost?options=--search_path%3D{POSTGRES_TEST_SCHEMA}",
+            **kwargs,
+        ),
+    )
+    await queue.connect()
+    return queue
+
+
+async def cleanup_queue(queue: Queue) -> None:
+    if isinstance(queue, RedisQueue):
+        await queue.redis.flushdb()
     await queue.disconnect()
+
+
+async def setup_postgres() -> None:
+    async with await psycopg.AsyncConnection.connect(
+        "postgres://postgres@localhost", autocommit=True
+    ) as conn:
+        await conn.execute(f"CREATE SCHEMA IF NOT EXISTS {POSTGRES_TEST_SCHEMA}")
+
+
+async def teardown_postgres() -> None:
+    async with await psycopg.AsyncConnection.connect(
+        "postgres://postgres@localhost", autocommit=True
+    ) as conn:
+        await conn.execute(f"DROP SCHEMA {POSTGRES_TEST_SCHEMA} CASCADE")
